backend/requirements.txt
flask
flask-sqlalchemy
flask-cors
werkzeug
ImageHash
Pillow
